♻️ refactor(WebLottoMachine): 함수형에서 클래스로 변경
처음 시작 시 금액 입력창에 포커스를 자바스크립트가 아닌 input tag autofocus로 변경

 비슷한 동작을 하는 기능끼리 묶어서 함수 구현
