+++ index.html
@@ -16,6 +16,7 @@
             <div class="buy-container">
               <input
                 required
+                autofocus
                 id="moneyAmount"
                 class="text-placeholder text-body"
                 placeholder="금액"
